+++ src/main/java/org/cesi/fablab/api/dto/SiteDTO.java
@@ -17,19 +17,19 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class SiteDTO {
-	private int id;
-	private String name;
-	private String description;
-	private List<HolidayEntity> holidaysList = new ArrayList<HolidayEntity>();
-	private List<RoomEntity> roomsList = new ArrayList<RoomEntity>();
+    private long id;
+    private String name;
+    private String description;
+    private List<HolidayEntity> holidaysList = new ArrayList<HolidayEntity>();
+    private List<RoomEntity> roomsList = new ArrayList<RoomEntity>();
 
-	public SiteDTO(SiteEntity siteEntity) {
-		super();
-		this.id = siteEntity.getId();
-		this.name = siteEntity.getName();
-		this.description = siteEntity.getDescription();
-		// this.holidaysList = siteEntity.getHolidaysList();
-		// this.roomsList = siteEntity.getRoomsList();
-	}
+    public SiteDTO(SiteEntity siteEntity) {
+        super();
+        this.id = siteEntity.getId();
+        this.name = siteEntity.getName();
+        this.description = siteEntity.getDescription();
+        // this.holidaysList = siteEntity.getHolidaysList();
+        // this.roomsList = siteEntity.getRoomsList();
+    }
 
 }
